Added info line and info column to retrieve parsing position information.  Made the interfaces for position information using a long for this. Tried to prevent concurrent use of a schema command - a schema command may be busy. Added book-keeping code to prevent seg fault, if a schema command is configured as validation command for a SAX parser or DOM building and the schema command is deleted on the way at script level.

+++ generic/domxslt.c
@@ -545,7 +545,7 @@ reportError (
     Tcl_DString dStr;
     char buffer[1024];
     const char *baseURI;
-    int  line, column;
+    long  line, column;
 
     Tcl_DStringInit (&dStr);
     baseURI = findBaseURI (node);
@@ -555,7 +555,7 @@ reportError (
     }
     if (node->nodeFlags & HAS_LINE_COLUMN) {
         domGetLineColumn (node, &line, &column);
-        sprintf (buffer, " at line %d, column %d:\n", line, column);
+        sprintf (buffer, " at line %ld, column %ld:\n", line, column);
         Tcl_DStringAppend (&dStr, buffer, -1);
         Tcl_DStringAppend (&dStr, str, -1);
     } else {
